refactor: move auth handlers into separated modules
